Fix item_efficiency data type from INTEGER to DECIMAL
🐛 Fixed data type mismatch:
- C++ code uses double for item_efficiency
- PostgreSQL schema incorrectly used INTEGER
- Updated to DECIMAL for proper precision

✅ Ensures data integrity during migration from SQLite

+++ database/schema/001_initial_schema.sql
@@ -151,7 +151,7 @@ CREATE TABLE companies (
     production_function INTEGER NOT NULL,
     production INTEGER NOT NULL,
     employees INTEGER NOT NULL,
-    item_efficiency INTEGER NOT NULL,
+    item_efficiency DECIMAL NOT NULL,
     capacity_vs_efficiency_split DECIMAL NOT NULL,
     recorded_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
